+++ renderer.php
@@ -196,6 +196,45 @@ class format_collapsibletopics_renderer extends format_section_renderer_base {
         $o = '';
         $sectionstyle = '';
 
+        if ($course->sectionprogress) {
+            $total = 0;
+            $complete = 0;
+            $cancomplete = isloggedin() && !isguestuser();
+            $modinfo = get_fast_modinfo($course);
+
+            $sectionmods = array();
+            $completioninfo = new completion_info($course);
+            if (!empty($modinfo->sections[$section->section])) {
+                foreach ($modinfo->sections[$section->section] as $cmid) {
+
+                    $thismod = $modinfo->cms[$cmid];
+
+                    if ($thismod->modname == 'label') {
+                        // Labels are special (not interesting for students)!
+                        continue;
+                    }
+
+                    if ($thismod->uservisible) {
+                        if (isset($sectionmods[$thismod->modname])) {
+                            $sectionmods[$thismod->modname]['name'] = $thismod->modplural;
+                            $sectionmods[$thismod->modname]['count']++;
+                        }
+                        else {
+                            $sectionmods[$thismod->modname]['name'] = $thismod->modfullname;
+                            $sectionmods[$thismod->modname]['count'] = 1;
+                        }
+                        if ($cancomplete && $completioninfo->is_enabled($thismod) != COMPLETION_TRACKING_NONE) {
+                            $total++;
+                            $completiondata = $completioninfo->get_data($thismod, true);
+                            if ($completiondata->completionstate == COMPLETION_COMPLETE || $completiondata->completionstate == COMPLETION_COMPLETE_PASS) {
+                                $complete++;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
         if ($section->section != 0) {
             // Only in the non-general sections.
             if (!$section->visible) {
@@ -233,6 +272,9 @@ class format_collapsibletopics_renderer extends format_section_renderer_base {
                 array('class' => 'sectionname'));
             // Add collapse toggle.
             if (course_get_format($course)->is_section_current($section)) {
+                if ($course->sectionprogress && $total > 0) {
+                    $o .= $this->section_progressbar($total, $complete);
+                }
                 $o .= '<a class="sectiontoggle' .
                     '" data-toggle="collapse" data-parent="accordion" href="#collapse-' .
                     $section->section .
@@ -240,6 +282,9 @@ class format_collapsibletopics_renderer extends format_section_renderer_base {
                     $section->section .
                     '">&nbsp;' . $sectionname . '</a> ';
             } else if ($section->section != 0) {
+                if ($course->sectionprogress && $total > 0) {
+                    $o .= $this->section_progressbar($total, $complete);
+                }
                 $o .= '<a class="sectiontoggle collapsed' .
                     '" data-toggle="collapse" data-parent="accordion" href="#collapse-' .
                     $section->section .
@@ -352,11 +397,19 @@ class format_collapsibletopics_renderer extends format_section_renderer_base {
      * @return string
      */
     protected function section_summary($section, $course, $mods) {
+        $activities = $this->section_activity_summary($section, $course, null);
+
         $o = '';
         $o .= html_writer::start_tag('div', array('class' => 'summarytext'));
         $o .= $this->format_summary_text($section);
         $o .= html_writer::end_tag('div');
-        $o .= $this->section_activity_summary($section, $course, null);
+        $o .= '<a class="sectiontoggle-bottom' .
+                    '" data-toggle="collapse" data-parent="accordion" href="#collapse-' .
+                    $section->section .
+                    '" aria-expanded="true" aria-controls="collapse-' .
+                    $section->section .
+                    '">' . $activities . '</a> ';
+
 
         return $o;
     }
@@ -427,5 +480,31 @@ class format_collapsibletopics_renderer extends format_section_renderer_base {
         } else {
             return array_merge($controls, $parentcontrols);
         }
+    }
+
+    protected function section_progressbar($total, $complete) {
+        $o = '';
+        $completion = new stdClass;
+        $completion->complete = $complete;
+        $completion->total = $total;
+        $percenttext = get_string('sectionprogresstext', 'format_collapsibletopics');
+        $percent = 0;
+        $current = 0;
+
+        if ($complete > 0) {
+            $current = (int)$complete;
+            $percent = (int)(($complete / $total) * 100);
+        }
+
+        $o .= '<div class="progress">';
+        $o .= '<div class="progress-bar progress-bar-info" role="progressbar" aria-valuenow="' . $current .'" ';
+        $o .= ' aria-valuemin="0" aria-valuemax="' . $total .'" style="width: ' . $percent . '%;">';
+        $o .= '<div class="progresstest">';
+        $o .= '<span class="sr-only">' . $percenttext . '</span>';
+        $o .= '</div>';
+        $o .= '</div>';
+        $o .= '</div>';
+
+        return $o;
     }
 }