+++ CHANGES.md
@@ -9,3 +9,6 @@ Changes
 * 2019-05-20 - Remove support for legacy themes.
 * 2019-08-20 - Add toggle icon for rtl languages.
 * 2019-08-20 - Fix issue #12 Wrong item labels in drop targets list.
+
+### Work in progress
+* 2019-08-21 - Add section progress bar feature (initiated by Chris Kenniburg)